chore: alternative post & calc functions
